+++ pages/index.js
@@ -1,16 +1,8 @@
-import {Box, styled} from "@mui/material";
 import NavBar from "@/public/components/NavBar";
 import {useUser} from "@auth0/nextjs-auth0/client";
 import Product from "@/pages/Product";
 import Intro from "@/pages/Intro";
-
-const OuterBox = styled(Box) ({
-    display: "flex",
-    flexDirection: "column",
-    borderRadius: "0.3125rem",
-    background: "#EEE",
-    height: "100vh",
-});
+import {OuterBox} from "@/public/components/common/Boxes";
 
 export default function Home() {
     const { user, error, isLoading } = useUser();
